structure: API to `__init__.py`
To make it easier to use `ratatoskr` the user API functions (decorators)
were moved to `__init.py`.

`protectron` decorator was moved to `__init__.py` and the past
`protectron` module was renamed to `args_parser.py` to contain the util
functions.

`register_operation` and `dispatch_event` were also moved to `__init.py`
in order to hide the internal implementation with `OperationRegistry`.

Signed-off-by: Gergő Nagy <gnagy@aiss.hu>