register_operation: refactor
* `register_operation` used the old way of declaring decorators that was
changed to `functools.wraps`.
* `register_operation` can now be called without parentheses using
`utils.doublewrap` decorator that handles this problem
* `args_parser` was renamed to `utils` to collect helper functions

* `register_operation` is able to get `class` or `instance` as a
parameter for `operation_wrapper`. This was neccesarry because python
will reuse the `opeation_wrapper` created on the first
`register_operation` decorator statement. Every time `register_operation` is called
a new `operation_wrapper` must be created. So it's partly a bugfix and a
feature.

Signed-off-by: Gergő Nagy <gnagy@aiss.hu>