operation_registry: `doc` option is added to `meta`
The `meta` field serves a purpose to hold control information for
ratatoskr about the event.

The newly introduced `doc` field is a Boolean type that controls if the
requestor of the event wants to get back the docstring of the registered
operation.

The base wrapper introduced a new class method `help(cls)` that returns
the docstring of the function. It can be reimplemented in derived
classes if the developer has intention to serve this documentation from
another source.

Signed-off-by: Gergő Nagy <contact@gergonagy.com>